# use eval() to improve CPU performance