@zhangting2020 2020-07-08
TODO：Once it is verified that the use of `eval` in GPU Kernel will not cause a significant increase in memory, we will apply it to CPU and GPU kernel.